backend/requirements_dev.txt
-r requirements.txt
appdirs==1.4.3
astroid==2.3.3
autopep8==1.5.2
bandit==1.6.2
black==19.10b0
dodgy==0.2.1
entrypoints==0.3
flake8==3.7.9
gitdb==4.0.4
GitPython==3.1.1
isort==4.3.21
jedi==0.15.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
parso==0.7.0
pathspec==0.8.0
pbr==5.4.5
pep8-naming==0.4.1
pluggy==0.13.1
prospector==1.2.0
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
pylama==7.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.1
python-jsonrpc-server==0.3.4
regex==2020.4.4
requirements-detector==0.6
rope==0.16.0
setoptconf==0.2.0
smmap==3.0.2
snowballstemmer==2.0.0
sqlalchemy-stubs==0.3
stevedore==1.32.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
ujson==1.35
wcwidth==0.1.9
wrapt==1.11.2
yapf==0.30.0
